requirements.txt
asgiref==3.7.2
Django==4.2.4
lxml==4.9.3
Pillow==10.0.0
python-pptx==0.6.21
sqlparse==0.4.4
typing_extensions==4.7.1
XlsxWriter==3.1.2
